+++ src/lib.rs
@@ -1,5 +1,6 @@
 #![allow(unused_variables)]
 #![allow(dead_code)]
+#![recursion_limit = "256"]
 #[macro_use]
 extern crate lazy_static;