+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.0
+**This version changed some mappings in presets, so the same input probably won't return the same result as version 1.0.0.**
+
+- Add more word mappings, so it's on par with mohan-cao's latest owoify.
+- Clean up codes by removing variable declarations in mappings.
+- Add uwuify and uvuify for convenience.
+
 ## 1.0.1
 - Bump up to 1.0.1